remove org2blog/wp-sourcecode-langs
org2blog/wp-sourcecode-langs duplicates a list of supported languages
which depends entirely on the version of the WordPress SyntaxHighlighter
being used, so it is impossible to maintain this list to be correct for
all users, and it is unrealistic to expect users to maintain it to be
accurate.

Additionally, the only effect the variable has is to force the shortcode
to be "text" in case an language is chosen which is not in this list.
In that case, forcing it to text is almost certainly not what the user
would want or expect.

Therefore it is better to adhere to the DRY rule and just eliminate this
duplication.