Delete annoying pop-up window
Not fully diagnosed yet, but org2blog/wp-upload-files-replace-urls
creates a new window somehow in the course of its main save-excursion
loop.  This patch deletes that window once it's no longer needed; though
if the *Org WP Export* buffer grabs an existing window, that window will
also be deleted. I'm not sure how to fix that problem.