Close 116: After publishing provide post or page metadata in hook fn
Added org2blog/wp-after-new-post-or-page-functions

Hook after posting a post or a page.

Added org2blog/wp-get-post-or-page

Gets the metaweblog entity for a post or page id. Made it interactive to
facilitate new features.

org2blog/wp-post-buffer
org2blog/wp-post-subtree

After a posting, get a copy of the post using the post id.

Incremented version to 1.0.1.

Added hook information to readme.